+++ NEWS.md
@@ -1,5 +1,8 @@
 # tibblify (development version)
 
+* New `parse_openapi_spec()` and `parse_openapi_schema()` to convert an
+  OpenAPI specification to a tibblify specification.
+
 * New `unpack_tspec()` to unpack the elements of `tib_row()` fields (#165).
 
 * Improved printing of lists parsed with `tspec_object()`.